# src/Dle.Control/openapi/dle-control.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Deep Link Engine — control plane",
    "description": "Link, domain, application, tenant and credential management, plus attribution and analytics reporting. Errors are RFC 9457 problem documents whose type member is one of the stable identifiers documented at https://docs.dle.dev/problems/. Writes accept an Idempotency-Key header: the same key with the same body replays the stored response, and the same key with a different body is a conflict.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/webhooks": {
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Lists the tenant's webhook subscriptions.",
        "description": "The shared secret is never returned; it is shown once, at creation.",
        "operationId": "ListWebhooks",
        "parameters": [
          {
            "name": "onlyActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebhookResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Registers a postback endpoint.",
        "description": "The destination is resolved and checked against the same target policy a link is checked against, so a webhook cannot be aimed at a private, loopback or cloud metadata address (T-02). The response carries the shared secret exactly once.",
        "operationId": "CreateWebhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/{id}": {
      "delete": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Stops delivering to a subscription.",
        "description": "Deactivation rather than deletion: the delivery history stays, because it is the only answer to a later question about what was and was not sent.",
        "operationId": "DeleteWebhook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/{id}/test": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Sends one signed test delivery and reports the result.",
        "description": "A real, fully signed delivery of type webhook.test through the same dispatcher the worker uses. Synchronous, so the integrator learns immediately whether the endpoint answered and what it answered.",
        "operationId": "TestWebhook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestWebhookResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWebhookRequest": {
        "required": [
          "url",
          "event_types"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Destination URL. Validated against the same target policy as a link, so a webhook\r\n            cannot be pointed at a cloud metadata endpoint (T-02)."
          },
          "event_types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Event types to deliver, for example `attribution.created`."
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the subscription is active on creation."
          }
        },
        "description": "Body of `POST /api/v1/webhooks` (FR-204)."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TestWebhookResponse": {
        "required": [
          "delivered",
          "outcome",
          "elapsed_ms",
          "payload"
        ],
        "type": "object",
        "properties": {
          "delivered": {
            "type": "boolean",
            "description": "Whether the endpoint accepted the delivery."
          },
          "outcome": {
            "type": "string",
            "description": "Outcome: `delivered`, `failed` or `unsendable`."
          },
          "response_code": {
            "type": [
              "null",
              "integer"
            ],
            "description": "HTTP status the endpoint returned, when it answered.",
            "format": "int32"
          },
          "error": {
            "type": [
              "null",
              "string"
            ],
            "description": "Short description of the failure, when there was one."
          },
          "elapsed_ms": {
            "type": "integer",
            "description": "How long the attempt took, in milliseconds.",
            "format": "int32"
          },
          "payload": {
            "type": "string",
            "description": "The exact body that was signed and sent, so the integrator can replay the check."
          },
          "signing_key_id": {
            "type": [
              "null",
              "string"
            ],
            "description": "Key identifier named in the `v2` slot of the signature."
          }
        },
        "description": "Result of `POST /api/v1/webhooks/{id}/test` (§B.7.3)."
      },
...
      "WebhookCreatedResponse": {
        "required": [
          "id",
          "url",
          "is_active",
          "created_at",
          "secret"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Subscription identifier.",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "description": "Destination URL."
          },
          "event_types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Subscribed event types."
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the subscription is active."
          },
          "created_at": {
            "type": "string",
            "description": "Creation instant.",
            "format": "date-time"
          },
          "secret": {
            "type": "string",
            "description": "The shared secret, base64 encoded, shown exactly once. Feed it to an HMAC-SHA-256 over\r\n`t + \".\" + body` to verify the `v1` member of `DLE-Signature`."
          },
          "signing_key_id": {
            "type": [
              "null",
              "string"
            ],
            "description": "Key identifier of the asymmetric slot at the moment of creation, so an integrator knows\r\nwhich key of `/.well-known/jwks.json` to look for while testing."
          }
        },
        "description": "Response to `POST /api/v1/webhooks`. The shared secret appears here and nowhere else\r\n(§B.5.2, §E.4.1 K4)."
      },
...
      "WebhookResponse": {
        "required": [
          "id",
          "url",
          "is_active",
          "created_at"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Subscription identifier.",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "description": "Destination URL."
          },
          "event_types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Subscribed event types."
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the subscription is active."
          },
          "created_at": {
            "type": "string",
            "description": "Creation instant.",
            "format": "date-time"
          }
        },
        "description": "Representation of a webhook subscription. The signing secret is never returned."
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "description": "A control-plane API key, presented as a bearer token or in the X-Api-Key header. The key looks like dle_<prefix>_<secret>; only its Argon2id hash is stored, so a lost key is replaced rather than recovered.",
        "scheme": "bearer"
      },
      "SdkKey": {
        "type": "apiKey",
        "description": "A key embedded in a customer application. It authenticates the two SDK ingestion routes and nothing else, and must be assumed extractable from the application binary.",
        "name": "X-Dle-Sdk-Key",
        "in": "header"
      }
    }
  },
...
}